fix: rename for kebab-case
