+++ src/components/gallery/hide-button.tsx
